- started to integrate in Tony's online update code.
I commented out some of it, just so that I could get it to compile in for tonight.  We need to add the missing attributes to MyMoneyEquity (that are in your MMQItem) so that all the functions will work correctly.  I hope you don't mind, I added parts of your MyMoneyQuoter class into the MyMoneyOnlinePriceUpdate class, because the latter class was already in the project, and it just made it easier.

-Kevin