* Added some new default reports, which are permutations of the old ones   * Modified the reports index list to show a tree view, organized by type of     report.   * Modified reports to remember the plain language date range setting,     so if you set a report to "Current Month", it will ALWAYS show the     current month.   * Added rudimentary help to report configuration dialog.   * Added 'close' button to report configuration dialog.
WARNING: There is a small possiblity of data loss with this checkin.  If you
had a customized report and it doesn't load correctly, DO NOT SAVE, or it will
wipe out the old report for good.  Instead, extract the <REPORTS> element from
the XML and send it me, and I will fix the problem asap.  I've tested this
extensively, so it shouldn't be a problem, but you never know.